chp14/django_ecommerce/requirements.txt
Django==1.8.2
django-embed-video==1.0.0
djangorestframework==3.3.2
funcsigs==0.4
mock==1.3.0
pbr==1.8.1
psycopg2==2.6.1
requests==2.9.0
six==1.10.0
stripe==1.9.2
wheel==0.24.0
